Refactor common util functions for /test
Prior to this commit, the uitl functions for e2e tests are distributed
over multiple test files in /test. And we had to tag specific test
files for unrelevant test suites that needs such util functions. For
example, init_test has the 'conformance' build tag on it, which is not
necessarily the case as conformance test suite does not need the
inti_test but needs the setup there.

This commit cleans up the util functions separated across different test
files and put them into /test/util so that we can tag the appropriate
test files for build only.
