+++ test/retry_test.go
@@ -26,10 +26,7 @@ import (
 	"time"
 
 	"github.com/tektoncd/pipeline/test/parse"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"knative.dev/pkg/apis"
-	duckv1 "knative.dev/pkg/apis/duck/v1"
 	knativetest "knative.dev/pkg/test"
 	"knative.dev/pkg/test/helpers"
 )
@@ -138,20 +135,4 @@ spec:
 			}
 		}
 	}
-}
-
-// This method is necessary because PipelineRunTaskRunStatus and TaskRunStatus
-// don't have an IsFailed method.
-func isFailed(t *testing.T, taskRunName string, conds duckv1.Conditions) bool {
-	t.Helper()
-	for _, c := range conds {
-		if c.Type == apis.ConditionSucceeded {
-			if c.Status != corev1.ConditionFalse {
-				t.Errorf("TaskRun status %q is not failed, got %q", taskRunName, c.Status)
-			}
-			return true
-		}
-	}
-	t.Errorf("TaskRun status %q had no Succeeded condition", taskRunName)
-	return false
 }
